Move Memo to many_types where it belongs
