+++ src/many-modules/Cargo.toml
@@ -30,4 +30,4 @@ proptest = "1.0.0"
 smol = "1.2.5"
 
 [features]
-memo-backward-compatible = []
+memo-backward-compatible = ["many-types/memo-backward-compatible"]
